fix: 아카이브 관련 권한 조절

+++ src/main/java/com/palettee/archive/service/ArchiveService.java
@@ -55,10 +55,10 @@ public class ArchiveService {
         return new ArchiveResponse(savedArchive.getId());
     }
 
-    public ArchiveListResponse getAllArchive(String color, String sort, Pageable pageable, User contextUser) {
+    public ArchiveListResponse getAllArchive(String color, String sort, Pageable pageable, User optionalUser) {
         Slice<Archive> archives = archiveRepository.findAllArchiveWithCondition(color, sort, pageable);
 
-        Long myId = contextUser == null ? 0L : contextUser.getId();
+        Long myId = getOptionalUserId(optionalUser);
 
         List<ArchiveSimpleResponse> list = archives
                 .map(it -> ArchiveSimpleResponse.toResponse(it, myId, likeRepository))
@@ -67,8 +67,12 @@ public class ArchiveService {
         return new ArchiveListResponse(list, SliceInfo.of(archives));
     }
 
-    public ArchiveListResponse getMainArchive(User contextUser) {
-        Long myId = contextUser == null ? 0L : contextUser.getId();
+    private Long getOptionalUserId(User optionalUser) {
+        return optionalUser == null ? 0L : optionalUser.getId();
+    }
+
+    public ArchiveListResponse getMainArchive(User optionalUser) {
+        Long myId = getOptionalUserId(optionalUser);
         List<ArchiveSimpleResponse> result = archiveRepository.getMainArchives().stream()
                 .map(it -> ArchiveSimpleResponse.toResponse(it, myId, likeRepository))
                 .toList();
@@ -93,28 +97,25 @@ public class ArchiveService {
     }
 
     public ArchiveListResponse getMyArchive(User user) {
-
-        Long myId = user == null ? 0L : user.getId();
-        List<ArchiveSimpleResponse> result = archiveRepository.getAllMyArchive(myId)
+        List<ArchiveSimpleResponse> result = archiveRepository.getAllMyArchive(user.getId())
                 .stream()
-                .map(it -> ArchiveSimpleResponse.toResponse(it, myId, likeRepository))
+                .map(it -> ArchiveSimpleResponse.toResponse(it, user.getId(), likeRepository))
                 .toList();
         return new ArchiveListResponse(result, null);
     }
 
     public ArchiveListResponse getLikeArchive(User user) {
-        Long myId = user == null ? 0L : user.getId();
-        List<Long> ids = likeRepository.findMyLikeList(myId);
+        List<Long> ids = likeRepository.findMyLikeList(user.getId());
         List<ArchiveSimpleResponse> result = archiveRepository.findAllInIds(ids)
                 .stream()
-                .map(it -> ArchiveSimpleResponse.toResponse(it, myId, likeRepository))
+                .map(it -> ArchiveSimpleResponse.toResponse(it, user.getId(), likeRepository))
                 .toList();
         return new ArchiveListResponse(result, null);
     }
 
-    public ArchiveListResponse searchArchive(String searchKeyword, Pageable pageable, User contextUser) {
+    public ArchiveListResponse searchArchive(String searchKeyword, Pageable pageable, User optionalUser) {
         List<Long> ids = tagRepository.findAllArchiveIds(searchKeyword);
-        Long myId = contextUser == null ? 0L : contextUser.getId();
+        Long myId = getOptionalUserId(optionalUser);
         Slice<Archive> archives = archiveRepository.searchArchive(searchKeyword, ids, pageable);
 
         List<ArchiveSimpleResponse> list = archives
@@ -125,13 +126,9 @@ public class ArchiveService {
     }
 
     @Transactional
-    public ArchiveResponse updateArchive(Long archiveId, ArchiveUpdateRequest archiveUpdateRequest, User contextUser) {
+    public ArchiveResponse updateArchive(Long archiveId, ArchiveUpdateRequest archiveUpdateRequest, User user) {
         Archive archive = getArchive(archiveId);
-
-        Long myId = contextUser == null ? 0L : contextUser.getId();
-        if (!archive.getUser().getId().equals(myId)) {
-            throw NotArchiveOwner.EXCEPTION;
-        }
+        checkArchiveOwner(user, archive);
 
         Archive updatedArchive = archive.update(archiveUpdateRequest);
 
@@ -143,8 +140,11 @@ public class ArchiveService {
     }
 
     @Transactional
-    public ArchiveResponse deleteArchive(Long archiveId) {
+    public ArchiveResponse deleteArchive(Long archiveId, User user) {
         Archive archive = getArchive(archiveId);
+
+        checkArchiveOwner(user, archive);
+
         tagRepository.deleteByArchive(archive);
         archiveImageRepository.deleteByArchive(archive);
         archiveRepository.delete(archive);
@@ -157,12 +157,19 @@ public class ArchiveService {
     }
 
     @Transactional
-    public void changeArchiveOrder(ChangeOrderRequest changeOrderRequest) {
+    public void changeArchiveOrder(ChangeOrderRequest changeOrderRequest, User user) {
         Map<Long, Integer> map = changeOrderRequest.orderRequest();
 
         for (Long pk : map.keySet()) {
             Archive archive = getArchive(pk);
+            checkArchiveOwner(user, archive);
             archive.updateOrder(map.get(pk));
+        }
+    }
+
+    private void checkArchiveOwner(User user, Archive archive) {
+        if (!archive.getUser().getId().equals(user.getId())) {
+            throw NotArchiveOwner.EXCEPTION;
         }
     }
